Store aelo_version into the datastore
